projects module: move setup logic and added assert cfg
